Move RefinementOrigin to separate file
